requirements.txt
scipy>=1.5.2
numpy>=1.19.1
pandas>=1.1.5
pyyaml>=5.3.1
pytorch>=1.7.1
future>=0.18.2
torchdiffeq>=0.2.0